+++ envisim_utils/src/indices.rs
@@ -40,6 +40,7 @@ impl Indices {
     /// ```
     /// # use envisim_utils::indices::*;
     /// let il = Indices::new(10);
+    /// assert!(il.is_empty());
     /// ```
     #[inline]
     pub fn new(capacity: usize) -> Self {
@@ -49,12 +50,14 @@ impl Indices {
         }
     }
 
-    /// Constructs a new `Indices`, filled with (0..length)
+    /// Constructs a new `Indices`, filled with units `(0..length)`
     ///
     /// # Examples
     /// ```
     /// # use envisim_utils::indices::*;
     /// let il = Indices::with_fill(10);
+    /// assert_eq!(il.first(), Some(9));
+    /// assert_eq!(il.last(), Some(0));
     /// ```
     #[inline]
     pub fn with_fill<NZ>(length: NZ) -> Self
@@ -64,9 +67,13 @@ impl Indices {
         let Ok(nz) = length.try_into() else {
             return Self::new(0);
         };
+        // By storing the list in reverse order, algos that need to be able to draw in order can
+        // always draw from the end, and a swap remove will guarantee that the selected units are
+        // stable, without incurring a higher cost needed in order to keep order in start of list.
         Indices {
-            list: (0..nz.get()).collect::<Vec<usize>>(),
+            list: (0..nz.get()).rev().collect::<Vec<usize>>(),
             indices: (0..nz.get())
+                .rev()
                 .map(|v| (v, v))
                 .collect::<FxHashMap<usize, usize>>(),
         }
@@ -78,8 +85,9 @@ impl Indices {
     /// ```
     /// # use envisim_utils::indices::*;
     /// let mut il = Indices::with_fill(10);
+    /// assert_eq!(il.len(), 10);
     /// il.clear();
-    /// assert_eq!(il.len(), 0);
+    /// assert!(il.is_empty());
     /// ```
     #[inline]
     pub fn clear(&mut self) {
@@ -93,7 +101,7 @@ impl Indices {
     /// ```
     /// # use envisim_utils::indices::*;
     /// let il = Indices::with_fill(4);
-    /// assert_eq!(il.list(), vec![0, 1, 2, 3]);
+    /// assert_eq!(il.list(), vec![3, 2, 1, 0]);
     /// ```
     #[must_use]
     #[inline]
@@ -106,7 +114,7 @@ impl Indices {
     /// # use envisim_utils::indices::*;
     /// let il = Indices::with_fill(4);
     /// let v: Vec<usize> = il.to_vec();
-    /// assert_eq!(il.list(), &v);
+    /// assert_eq!(v, vec![3, 2, 1, 0]);
     /// ```
     #[must_use]
     #[inline]
@@ -118,7 +126,7 @@ impl Indices {
     /// ```
     /// # use envisim_utils::indices::*;
     /// let il = Indices::with_fill(4);
-    /// assert_eq!(il.get(3).unwrap(), 3);
+    /// assert_eq!(il.get(3), Some(0));
     /// assert_eq!(il.get(10), None);
     /// ```
     #[must_use]
@@ -130,8 +138,10 @@ impl Indices {
     /// # Examples
     /// ```
     /// # use envisim_utils::indices::*;
-    /// let il = Indices::with_fill(4);
-    /// assert_eq!(il.first().unwrap(), 0);
+    /// let mut il = Indices::with_fill(4);
+    /// assert_eq!(il.first(), Some(3));
+    /// il.clear();
+    /// assert_eq!(il.first(), None);
     /// ```
     #[must_use]
     #[inline]
@@ -142,8 +152,10 @@ impl Indices {
     /// # Examples
     /// ```
     /// # use envisim_utils::indices::*;
-    /// let il = Indices::with_fill(4);
-    /// assert_eq!(il.last().unwrap(), 3);
+    /// let mut il = Indices::with_fill(4);
+    /// assert_eq!(il.last(), Some(0));
+    /// il.clear();
+    /// assert_eq!(il.last(), None);
     /// ```
     #[must_use]
     #[inline]
@@ -154,7 +166,7 @@ impl Indices {
     /// # Examples
     /// ```
     /// # use envisim_utils::indices::*;
-    /// # use envisim_utils::random::*;
+    /// use envisim_utils::random::*;
     /// let il = Indices::with_fill(4);
     /// let mut rng = SmallRng::seed_from_u64(4242);
     /// assert!(il.draw(&mut rng).is_some());
@@ -168,23 +180,6 @@ impl Indices {
         R: RandomNumberGenerator,
     {
         rng.relement(&self.list).copied()
-    }
-
-    /// Returns the next sequential unit after `from`, not including itself, if it exists.
-    ///
-    /// # Examples
-    /// ```
-    /// # use envisim_utils::indices::*;
-    /// let mut il = Indices::with_fill(4);
-    /// assert_eq!(il.seq_after(0, 4), Some(1));
-    /// assert_eq!(il.seq_after(3, 4), None);
-    /// il.remove(2);
-    /// assert_eq!(il.seq_after(1, 4), Some(3));
-    /// ```
-    #[must_use]
-    #[inline]
-    pub fn seq_after(&self, from: usize, max: usize) -> Option<usize> {
-        ((from + 1)..max).find(|&id| self.contains(id))
     }
 
     /// Checks if the list contains an index
@@ -342,6 +337,6 @@ mod tests {
     #[test]
     fn with_fill() {
         let il = Indices::with_fill(4);
-        assert_eq!(il.list(), vec![0, 1, 2, 3]);
+        assert_eq!(il.list(), vec![3, 2, 1, 0]);
     }
 }